Move button from storage to worker capacity.  Update text
